[billy-olsen,r=] Make post-deploy-config more idempotent.
The bin/post-deploy-config script will add a port and change the
neutron-gateway config each time it is called. This change modifies
the script to first attempt to use existing attached ports instead
of always attaching a new one. In order to do this, the script must
skip over any ports which are being used for juju unit.

Included in this change is a small change to the quota scripts which
will use the openstack-client instead of the keystone client in order
to avoid deprecation notices.