Use correct scope for authentication in post deployment script
Post deployment configuration script currently relies on
the OS_DOMAIN_NAME environment variable and requests a
domain scoped token when authenticating to Keystone.

Holding a domain scoped token is generally not useful for
actually using resources on a cloud and in most if not all
circumstances you should use a project scoped token.

When we initially implemented API v3 support in the Keystone
charm [0] we chose to use a advanced policy that enables
domain segregation and administration delegation [1]. It is
important to understand that this differs from the default
policy distributed with the upstream Keystone project.

In the initial implementation it was necessary to switch
between using a `domain` scoped and `project` scoped token to
successfully manage a cloud, but since the introduction of
`is_admin` functionality in Keystone [2][3][4] and our
subsequent adoption of it in Keystone charm [5], this is no
longer necessary.

0: https://github.com/openstack/charm-keystone/commit/c283a1c922a903133ef31627203a56b951129324
1: https://github.com/openstack/keystone/commit/c7a5c6cf27a80ca50db9f1a1a74e8795eeefd9d1
2: https://github.com/openstack/keystone/commit/e7023697a884759716d0a01605825a3af90d4db6
3: https://github.com/openstack/keystone/commit/e923a14afdd8994129a17ece57dd220c307888dc
4: https://github.com/openstack/keystone/commit/9804081a80ef815a86407a64f967986a7bf9ba25
5: https://github.com/openstack/charm-keystone/commit/10e3d84effcadd9c527d5d18f8bdd0d25003b85a