Remove superflous call to juju run in openrc frontend
The frontend script determines which Keystone API version
to use based on Keystone application version and a call to
`juju config`.  Based on this information it sources the
appropriate backend script.

The backend script does the `juju run` call to get unit
private address.

The fact that the frontend script currently also does this
call without a need for the information is a bug.

This commit addresses that.