TFix random admin project selection bug
When the neutron scripts workout the admin project id they iterate
over all the projects with a name that matches, when a match is found
they exit. However, two admin projects are created and which
order the script finds them in is probably down to their randomly
allocated uuid meaning networks may be created in the wrong admin
project.

The keystone session object has already selected the
correct project based on environment variables so extract the project
id from there.