Configure highest available volume api for tempest
Ensure the volumev3 endpoint is used if it is available, and volumev2
is used otherwise.

Tempest will run volume tests against one volume API, either volumev2
or volumev3. As of Pike, volumev3 is enabled by the cinder charm.

The deprecated api_v1, api_v2, and api_v3 options are dropped in favor
of catalog_type, which defaults to volumev3.