Stop processing units after first match
Starting Juju 2.5 the output of `juju status` with
format modifier contains information about storage
and volumes used in the model when present.

This makes the `openrc` scripts fail due to the awk
program returning multiple matches from its search
for the `keystone` unit.

Applications appear first in the output, so modify
the awk script to stop processing after first match
to amend the situation.